enhance user page and register
Refactor UserFormPage and UserPage for improved UI

Restructured UserFormPage.xaml to use ScrollView and
VerticalStackLayout for better scrolling and accessibility.
Added Labels for Entry fields and styled buttons with
new colors and attributes.

Updated UserPage.xaml to also utilize ScrollView and
VerticalStackLayout. Repositioned and styled the Registrar
button, replaced Frame with Border in CollectionView,
and organized user information layout for a cleaner look.